Add Validation and Logging

+++ src/services/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 
 import { IdeaEntity } from '../../entity/idea.entity';
@@ -7,7 +7,10 @@ import { IdeaDTO } from 'src/DTO/idea.dto';
 
 @Injectable()
 export class IdeaService {
-  constructor(@InjectRepository(IdeaEntity) private ideaRepository: Repository<IdeaEntity>) { }
+  constructor(
+    @InjectRepository(IdeaEntity)
+    private ideaRepository: Repository<IdeaEntity>,
+  ) {}
 
   async showAll() {
     return await this.ideaRepository.find();
@@ -20,16 +23,29 @@ export class IdeaService {
   }
 
   async showOne(id) {
-    return await this.ideaRepository.findOne({ where: { id } });
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    return idea;
   }
 
   async update(id, data: Partial<IdeaDTO>) {
+    let idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.update({ id }, data);
-    return await this.ideaRepository.findOne({id});
+    idea = await this.ideaRepository.findOne({ where: { id } });
+    return idea;
   }
 
   async destroy(id) {
-    await this.ideaRepository.delete({id});
-    return { deleted: true };
+    const idea = await this.ideaRepository.findOne({ where: { id } });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    await this.ideaRepository.delete({ id });
+    return idea;
   }
 }
